Frontend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.1.0
Automat==24.8.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bitarray==2.9.3
bitstring==4.2.3
blinker==1.9.0
brave-search==0.1.8
cap_from_youtube==0.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorlog==6.9.0
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
Cython==3.0.12
defusedxml==0.7.1
distro==1.9.0
enum34==1.1.10
filelock==3.18.0
Flask==3.1.1
flatbuffers==24.12.23
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2025.3.0
funcsigs==1.0.2
future==1.0.0
fuzz==0.1.1
fuzzywuzzy==0.18.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.25.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
ject==0.0.3
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
namex==0.0.8
nba_api==1.9.0
networkx==3.4.2
numpy==1.26.4
openai==1.65.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pafy==0.5.5
pandas==2.2.3
pandas-stubs==2.2.3.241126
paramiko==3.5.1
parsel==1.10.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.0.0
pinecone==6.0.1
pinecone-client==6.0.0
pinecone-plugin-assistant==1.3.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.0
Protego==0.4.0
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pygame==2.5.2
Pygments==2.19.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
pypdfium2==4.30.1
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
scipy==1.15.2
scp==0.15.0
Scrapy==2.12.0
seaborn==0.13.2
service-identity==24.2.0
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
tldextract==5.1.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
Twisted==24.11.0
types-beautifulsoup4==4.12.0.20250204
types-html5lib==1.1.11.20241018
types-pytz==2025.1.0.20250204
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2024.2
ulog==0.3.3
ultralytics==8.3.98
ultralytics-thop==2.0.14
urllib3==2.3.0
vidgear==0.3.3
w3lib==2.3.1
watchdog==4.0.1
way==0.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
yt-dlp==2025.3.27
zope.interface==7.2
